requirements.txt
appdirs==1.4.3
astroid==2.1.0
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
google==2.0.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
newsapi-python==0.2.6
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.2.2
python-telegram-bot==12.4.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.12.0
soupsieve==2.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.5
wrapt==1.11.1
zipp==3.0.0
